requirements.txt
python-telegram-bot
staticmap
googletrans
iexfinance
matplotlib
